app/requirements.txt
django~=3.1
djangorestframework==3.12.4
django-tastypie==0.14.3